+++ Extension/src/background/api/network/settings.ts
@@ -40,6 +40,38 @@ export class NetworkSettings {
     /**
      * Default base url for downloading filter rules.
      *
+     * MV3_REMOTE_POLICY.
+     * This keyword can be used to grep all code related to MV3 remote
+     * hosting policy.
+     *
+     * In MV3 extension we download a so-called "Quick Fixes filter" which
+     * is used for fixing major issues without the need to update the
+     * extension or custom filter lists added by the users themselves.
+     * Having this logic is particularly important for an ad blocker since
+     * websites breakages can occur at any time and we need to be able to
+     * fix them ASAP.
+     *
+     * We make sure that all these rules that come from the filter
+     * were in compliance with CWS policies:
+     * "Fetching a remote configuration file for A/B testing or determining
+     * enabled features, where all logic for the functionality is contained
+     * within the extension package".
+     *
+     * 1. Network rules from the Quick Fixes filter is converted to DNR
+     *    rules and applied via dynamic rules.
+     * 2. Cosmetic rules are interpreted in the code. For example, hiding
+     *    elements OR on the contrary, unhiding them when it is necessary.
+     *    At the same time the cosmetic rules logic is contained in the
+     *    extension package.
+     *
+     * Quick Fixes filter contents can be examined here:
+     * https://filters.adtidy.org/extension/chromium-mv3/filters/24.txt.
+     *
+     * To ensure compliance with Chrome Store policies, we have safeguards
+     * that restrict execution to rules that are included into the extension
+     * package and can be reviewed there. These safeguards can be found by
+     * searching for 'JS_RULES_EXECUTION'.
+     *
      * @private
      */
     private readonly DEFAULT_FILTER_RULES_BASE_URL = 'https://filters.adtidy.org/extension';
@@ -100,11 +132,42 @@ export class NetworkSettings {
     /**
      * Used to set the base url for filter rules through the local storage for testing purposes.
      *
+     * MV3_REMOTE_POLICY.
+     * This keyword can be used to grep all code related to MV3 remote
+     * hosting policy.
+     *
+     * In MV3 extension we download a so-called "Quick Fixes filter" which
+     * is used for fixing major issues without the need to update the
+     * extension or custom filter lists added by the users themselves.
+     * Having this logic is particularly important for an ad blocker since
+     * websites breakages can occur at any time and we need to be able to
+     * fix them ASAP.
+     *
+     * We make sure that all these rules that come from the filter
+     * were in compliance with CWS policies:
+     * "Fetching a remote configuration file for A/B testing or determining
+     * enabled features, where all logic for the functionality is contained
+     * within the extension package".
+     *
+     * 1. Network rules from the Quick Fixes filter is converted to DNR
+     *    rules and applied via dynamic rules.
+     * 2. Cosmetic rules are interpreted in the code. For example, hiding
+     *    elements OR on the contrary, unhiding them when it is necessary.
+     *    At the same time the cosmetic rules logic is contained in the
+     *    extension package.
+     *
+     * Quick Fixes filter contents can be examined here:
+     * https://filters.adtidy.org/extension/chromium-mv3/filters/24.txt.
+     *
+     * To ensure compliance with Chrome Store policies, we have safeguards
+     * that restrict execution to rules that are included into the extension
+     * package and can be reviewed there. These safeguards can be found by
+     * searching for 'JS_RULES_EXECUTION'.
+     *
      * @returns The base url for filter rules.
      */
     private async getFilterRulesBaseUrl(): Promise<string> {
-        // We don't need to set base url in MV3 because we cannot update filters
-        // via patches.
+        // We don't need to set base url in MV3 because we cannot update filters via patches.
         // TODO: Remove check when filters will support patches in MV3.
         if (!__IS_MV3__) {
             const url = localStorage.getItem(this.FILTERS_BASE_URL_KEY);
@@ -126,6 +189,39 @@ export class NetworkSettings {
         // first of all check whether it is mv3-build
         // https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2985
         if (__IS_MV3__) {
+            /**
+             * MV3_REMOTE_POLICY.
+             * This keyword can be used to grep all code related to MV3 remote
+             * hosting policy.
+             *
+             * In MV3 extension we download a so-called "Quick Fixes filter" which
+             * is used for fixing major issues without the need to update the
+             * extension or custom filter lists added by the users themselves.
+             * Having this logic is particularly important for an ad blocker since
+             * websites breakages can occur at any time and we need to be able to
+             * fix them ASAP.
+             *
+             * We make sure that all these rules that come from the filter
+             * were in compliance with CWS policies:
+             * "Fetching a remote configuration file for A/B testing or determining
+             * enabled features, where all logic for the functionality is contained
+             * within the extension package".
+             *
+             * 1. Network rules from the Quick Fixes filter is converted to DNR
+             *    rules and applied via dynamic rules.
+             * 2. Cosmetic rules are interpreted in the code. For example, hiding
+             *    elements OR on the contrary, unhiding them when it is necessary.
+             *    At the same time the cosmetic rules logic is contained in the
+             *    extension package.
+             *
+             * Quick Fixes filter contents can be examined here:
+             * https://filters.adtidy.org/extension/chromium-mv3/filters/24.txt.
+             *
+             * To ensure compliance with Chrome Store policies, we have safeguards
+             * that restrict execution to rules that are included into the extension
+             * package and can be reviewed there. These safeguards can be found by
+             * searching for 'JS_RULES_EXECUTION'.
+             */
             return `${this.filtersRulesBaseUrl}/chromium-mv3`;
         }